refactor(rvr): update rvr to have 64bit PC (#2923)
update rvr to have 64bit PC. This PR consist of multiple parts:
1) Updating Intermediate Representations to have 64bit PC instead of
32bit
2) Updating Lifter to use 64bit PC
3) Updating Emitter to use 64bit PC for its function arguments, update
PC related constants to 64bit.
4) Adding boundary checks during compile-time for statically computable
PCs (`JAL` and branching instructions)

resolves INT-8319

+++ crates/rvr/rvr-openvm/c/block/openvm_block_metered_segment.h
@@ -13,7 +13,7 @@ typedef struct MeteredSegmentCheckpointResult {
 } MeteredSegmentCheckpointResult;
 
 static __attribute__((always_inline)) inline uint8_t begin_block(
-    RvState* restrict state, uint32_t pc, uint32_t block_insn_count) {
+    RvState* restrict state, uint64_t pc, uint32_t block_insn_count) {
   return trace_block_with_segment_check(state, pc, block_insn_count);
 }
